dynamic report的数据transform

+++ springboot3-report/springboot3-ireport-demo/src/main/java/org/vivi/framework/ireport/demo/model/dataset/ReportDataSet.java
@@ -11,7 +11,7 @@ public class ReportDataSet {
 
     private Integer rtId;
 
-    private Integer sheetIndex;
+    private Long sheetIndex;
 
     private String rtSql;
